userspace: split scheduler-activation predicates out of policies.go
Move the scheduler-activation predicate trio (policyRuleInactive,
PolicyInactive, PolicyInactiveFn) verbatim from policies.go into a new
policies_scheduler.go. These three form the SSOT predicate shared between
the snapshot builder and the read-only show surfaces (#3062/#3104/#3414);
grouping them keeps that concern in one file.

Pure code-motion, no logic change: the declarations are byte-identical
slices of the origin/master policies.go. Refs #4421 (codex-173 #4).
